requirements.txt

Flask==2.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
